No real need to extract this
